Do not download if files are available

+++ scripts/download.sh
@@ -3,9 +3,16 @@
 # author oleg.shpynov@jetbrains.com
 
 echo "Downloading ENCODE data"
-wget -r ftp://ftp-trace.ncbi.nlm.nih.gov/sra/sra-instant/reads/ByRun/sra/SRR/SRR568/SRR568364/SRR568364.sra -O SRR568364.sra
-wget -r ftp://ftp-trace.ncbi.nlm.nih.gov/sra/sra-instant/reads/ByRun/sra/SRR/SRR568/SRR568365/SRR568365.sra -O SRR568365.sra
-
+if [ ! -f SRR568364.sra ]; then
+    wget -r ftp://ftp-trace.ncbi.nlm.nih.gov/sra/sra-instant/reads/ByRun/sra/SRR/SRR568/SRR568364/SRR568364.sra -O SRR568364.sra
+fi
+if [ ! -f SRR568365.sra ]; then
+    wget -r ftp://ftp-trace.ncbi.nlm.nih.gov/sra/sra-instant/reads/ByRun/sra/SRR/SRR568/SRR568365/SRR568365.sra -O SRR568365.sra
+fi
 echo "Downloading Roadmapepigenomics data"
-wget -r ftp://ftp-trace.ncbi.nlm.nih.gov/sra/sra-instant/reads/ByRun/sra/SRR/SRR787/SRR787515/SRR787515.sra -O SRR787515.sra
-wget -r ftp://ftp-trace.ncbi.nlm.nih.gov/sra/sra-instant/reads/ByRun/sra/SRR/SRR787/SRR787516/SRR787516.sra -O SRR787516.sra
+if [ ! -f SRR787515.sra ]; then
+    wget -r ftp://ftp-trace.ncbi.nlm.nih.gov/sra/sra-instant/reads/ByRun/sra/SRR/SRR787/SRR787515/SRR787515.sra -O SRR787515.sra
+fi
+if [ ! -f SRR787516.sra ]; then
+    wget -r ftp://ftp-trace.ncbi.nlm.nih.gov/sra/sra-instant/reads/ByRun/sra/SRR/SRR787/SRR787516/SRR787516.sra -O SRR787516.sra
+fi
